+++ packages/flutter_pptx/CHANGELOG.md
@@ -1,3 +1,4 @@
-## 0.0.1
+## 0.1.0
 
-* TODO: Describe initial release.
+- Adding support for widget slides
+- Added support for asset images